Move API calls inside component
